# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comfort taxi API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/driverFiredLog": {
            "get": {
                "tags": [
                    "driverFiredLog"
                ],
                "summary": "Get list of driverFiredLog",
                "operationId": "ed54d6f26ec59c64661c59034429fe04",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "400": {
                        "description": "driverFiredLog isn't found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/driverFiredLog/{id}": {
            "get": {
                "tags": [
                    "driverFiredLog"
                ],
                "summary": "Get driverFiredLog by id",
                "description": "Get driverFiredLog by id",
                "operationId": "getdriverFiredLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of driverFiredLog class",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "driverFiredLog": {
                                            "$ref": "#/components/schemas/DriverFiredLogResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "driverFiredLog is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "driverFiredLog is not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "driverFiredLog"
                ],
                "summary": "Create driverFiredLog",
                "description": "Create new driverFiredLog",
                "operationId": "createdriverFiredLog",
                "requestBody": {
                    "description": "Send data to create a new driverFiredLog",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "$ref": "#/components/schemas/DriverFiredLogCreateRequest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "task": {
                                            "$ref": "#/components/schemas/DriverFiredLogResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error when creating driverFiredLog",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error when creating driverFiredLog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "driverFiredLog"
                ],
                "summary": "Delete driverFiredLog by id",
                "description": "Delete driverFiredLog by id",
                "operationId": "deletedriverFiredLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of driverFiredLog",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "driverFiredLog": {
                                            "$ref": "#/components/schemas/DriverFiredLogResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "driverFiredLog is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "driverFiredLog class is not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "driverFiredLog"
                ],
                "summary": "Update driverFiredLog",
                "description": "Update driverFiredLog",
                "operationId": "updatedriverFiredLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of driverFiredLog",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Send data to update a driverFiredLog",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "driverFiredLog": {
                                        "$ref": "#/components/schemas/DriverFiredLogUpdateRequest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "driverFiredLog": {
                                            "$ref": "#/components/schemas/DriverFiredLogResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "driverFiredLog is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "driverFiredLog is not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error when updating driverFiredLog",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error when updating driverFiredLog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DriverFiredLogCreateRequest": {
                "required": [
                    "driver_id",
                    "date_hired",
                    "date_fired"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "date_hired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "date_fired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "driver_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DriverFiredLogCreateRequest"
                }
            },
            "DriverFiredLogUpdateRequest": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "date_hired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "date_fired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "driver_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DriverFiredLogUpdateRequest"
                }
            },
...
            "DriverFiredLogResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "date_hired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "date_fired": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "driver_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DriverFiredLog"
                }
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}